requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
pandas==2.0.3
undetected-chromedriver==3.5.3
requests==2.31.0
proxy-rotation==0.9.0
TwoCaptcha==0.0.1
PyJWT==2.8.0
transformers==4.33.2
torch==2.0.1
python-multipart==0.0.6
aiofiles==23.1.0
beautifulsoup4==4.12.2
